Use Table component.
