+++ packages/components/src/components/Table.styles.tsx
@@ -1,5 +1,21 @@
 import tw, { css, styled } from 'twin.macro';
 
+export const Table = styled.table(() => [tw`w-full border-collapse`]);
+
+export const TableBody = styled.tbody(() => []);
+
+export const TableHeader = styled.thead(() => [
+  css`
+    th {
+      ${tw`px-2 font-semibold text-left text-xs text-gray-300 dark:text-gray-600 uppercase whitespace-nowrap`}
+    }
+
+    &:n-th-of-type(3) {
+      ${tw`hidden md:table-cell`}
+    }
+  `
+]);
+
 export const TableItem = styled.tr(() => [
   tw`font-medium text-xs text-gray-500 dark:text-gray-400 border-0 border-b border-solid border-gray-300 dark:border-gray-800 last:border-0`,
   css`
@@ -74,3 +90,33 @@ export const TableItemButton = styled.button(() => [
 ]);
 
 export const TableItemDate = styled.span(() => [tw`md:whitespace-nowrap`]);
+
+export const TablePagination = styled.div(() => [
+  css`
+    ${tw`flex justify-center w-full`}
+
+    ul {
+      ${tw`flex list-none`}
+    }
+
+    li {
+      ${[
+        tw`inline-block mx-1 text-sm bg-gray-200 dark:(text-gray-300 bg-gray-700) rounded-lg select-none`,
+        tw`hover:opacity-70 transition duration-200 ease-in-out`,
+      ]}
+
+      &.selected {
+        ${tw`text-white bg-black dark:(text-gray-900 bg-gray-100)`}
+      }
+
+      &.next,
+      &.previous {
+        ${tw`hidden`}
+      }
+    }
+
+    a {
+      ${tw`flex items-center justify-center h-8 w-8 cursor-pointer`}
+    }
+  `,
+]);
